Working through the 'fast foward closure'.
I still have a few things left to do, but it looks like we have the same answers for alibaba.  Note that I added the RDFS MT axioms to the full forward closure, and that slowed it down quite a bit.  I have not checked either the counts or the specific statements generated against rdf-generic nor have I tested other datasets.  The things that I have left to do are the 3-part rules that are part of their custom program (steps 11 and 13).  This puts the time to load and close alibaba at 5219ms.

I have not looked at back-chaining rdfs:Resource and owl:sameAs.

I am not maintaining an explicit/inferred/axiom flag on the statements.

I am not writing out the proofs.

I have not done TM.

fast:
Computed closure in 2735ms yeilding 91068 statements total, 45416 inferences, entailmentsPerSec=16605
termId: #entries=18299, height=3, #nodes=72, #leaves=1143, #(nodes+leaves)=1215, #distinctOnQueue=454, queueCapacity=500
idTerm: #entries=18299, height=3, #nodes=72, #leaves=1143, #(nodes+leaves)=1215, #distinctOnQueue=441, queueCapacity=500
spo: #entries=91068, height=3, #nodes=195, #leaves=3327, #(nodes+leaves)=3522, #distinctOnQueue=452, queueCapacity=500
pos: #entries=91068, height=3, #nodes=350, #leaves=5686, #(nodes+leaves)=6036, #distinctOnQueue=453, queueCapacity=500
osp: #entries=91068, height=3, #nodes=294, #leaves=4798, #(nodes+leaves)=5092, #distinctOnQueue=444, queueCapacity=500

full:
Computed closure of 12 rules in 3 rounds and 12313ms yeilding 91068 statements total, 45275 inferences, entailmentsPerSec=3677
termId: #entries=18299, height=3, #nodes=72, #leaves=1143, #(nodes+leaves)=1215, #distinctOnQueue=454, queueCapacity=500
idTerm: #entries=18299, height=3, #nodes=72, #leaves=1143, #(nodes+leaves)=1215, #distinctOnQueue=454, queueCapacity=500
spo: #entries=91068, height=3, #nodes=195, #leaves=3328, #(nodes+leaves)=3523, #distinctOnQueue=452, queueCapacity=500
pos: #entries=91068, height=3, #nodes=350, #leaves=5687, #(nodes+leaves)=6037, #distinctOnQueue=496, queueCapacity=500
osp: #entries=91068, height=3, #nodes=295, #leaves=4820, #(nodes+leaves)=5115, #distinctOnQueue=454, queueCapacity=500


git-svn-id: svn+ssh://svn.code.sf.net/p/bigdata/code/trunk@268 8f7bc0f5-282e-0410-95e3-8d296e9bb460

+++ bigdata-rdf/src/java/com/bigdata/rdf/model/StatementEnum.java
@@ -58,29 +58,29 @@ public enum StatementEnum {
     /**
      * Something that is directly entailed by the appropriate model theory.
      */
-    Axiom(0),
+    Axiom((byte)0),
     /**
      * A statement that was inserted into the database explicitly by the
      * application.
      */
-    Explicit(1),
+    Explicit((byte)1),
     /**
      * A statement that was inferred from the explicit statements by the
      * appropriate model theory.
      */
-    Inferred(2),
+    Inferred((byte)2),
     /**
      * A statement that is undergoing truth maintenance.
      */
-    Suspended(3);
+    Suspended((byte)3);
 
-    private final int code;
+    private final byte code;
     
-    private StatementEnum(int code) {
+    private StatementEnum(byte code) {
         this.code = code;
     }
 
-    public int code() {
+    public byte code() {
         
         return code;
